HackStudio: Moved FormEditor related logic to a separate folder
